Filter Errors Before Raising Exception
When `Capybara::Chromedriver::Logger.raise_js_errors` is turned on,
errors will be output and thrown in the tests even if the text for these
errors are filtered by one of the regular expressions in the
`Capybara::Chromedriver::Logger.filters` collection. Ensure that the
error does not match one of these filters before proceeding to log it as
an error (and throw an exception in test teardown).